microservices/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
# MongoDB
motor==3.3.2
pymongo==4.6.1
# Legacy SQL support (can be removed if fully migrated)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
# Auth
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
pyjwt==2.10.0
# Google OAuth
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
# HTTP client
httpx==0.25.1
aiofiles==23.2.1
# Redis
redis==5.0.1
# Monitoring
prometheus-client==0.19.0
# Utilities
python-dateutil==2.8.2
requests==2.31.0
cryptography==41.0.7

